requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-cloud-aiplatform==1.36.1
google-generativeai==0.3.0
requests==2.31.0
beautifulsoup4==4.12.2
pillow==10.0.1
python-dotenv==1.0.0
textblob==0.17.1
nltk==3.8.1
newspaper3k==0.2.8
validators==0.22.0
werkzeug==2.3.7
gunicorn==21.2.0
spacy==3.7.2
scikit-learn==1.3.0
numpy==1.24.3
aiohttp==3.8.5
asyncio-throttle==1.0.2
textstat==0.7.3
opencv-python==4.8.0.76
